derive endpoints from dns, move cli e2e tests to tests/
