+++ topk-cli/src/commands/dataset.rs
@@ -138,7 +138,7 @@ impl fmt::Display for ListDatasetsResult {
 #[derive(Serialize, Deserialize)]
 #[serde(transparent)]
 pub struct GetDatasetResult {
-    pub(crate) dataset: Dataset,
+    pub dataset: Dataset,
 }
 
 impl From<DatasetPb> for GetDatasetResult {
@@ -169,7 +169,7 @@ impl fmt::Display for GetDatasetResult {
 #[derive(Serialize, Deserialize)]
 #[serde(transparent)]
 pub struct CreateDatasetResult {
-    pub(crate) dataset: Dataset,
+    pub dataset: Dataset,
 }
 
 impl From<DatasetPb> for CreateDatasetResult {
@@ -189,7 +189,7 @@ impl fmt::Display for CreateDatasetResult {
 #[derive(Serialize, Deserialize)]
 #[serde(transparent)]
 pub struct UpdateDatasetResult {
-    pub(crate) dataset: Dataset,
+    pub dataset: Dataset,
 }
 
 impl From<DatasetPb> for UpdateDatasetResult {
@@ -272,208 +272,4 @@ pub async fn delete<C: DatasetsClient>(
     client.delete(&args.dataset).await?;
 
     Ok(DeleteDatasetResult { deleted: true })
-}
-
-#[cfg(test)]
-mod tests {
-    use super::{
-        CreateDatasetResult, Dataset, DeleteDatasetResult, GetDatasetResult, UpdateDatasetResult,
-    };
-    use crate::commands::test_context::{CliTestContext, OutputJsonExt};
-    use assert_cmd::Command;
-    use test_context::test_context;
-
-    fn cmd() -> Command {
-        Command::cargo_bin("topk").unwrap()
-    }
-
-    #[test_context(CliTestContext)]
-    #[tokio::test]
-    async fn list(ctx: &mut CliTestContext) {
-        let name = ctx.wrap("test");
-        cmd()
-            .args(["dataset", "create", "--region", &ctx.region, &name])
-            .output()
-            .unwrap();
-
-        let out = cmd()
-            .args(["-o", "json", "dataset", "list"])
-            .output()
-            .unwrap();
-        assert!(
-            out.status.success(),
-            "{}",
-            String::from_utf8_lossy(&out.stderr)
-        );
-        let datasets: Vec<Dataset> = out.json_lines().unwrap();
-        let names: Vec<&str> = datasets.iter().map(|d| d.name.as_str()).collect();
-        assert!(
-            names.contains(&name.as_str()),
-            "created dataset not in list: {:?}",
-            names
-        );
-    }
-
-    #[test_context(CliTestContext)]
-    #[tokio::test]
-    async fn create(ctx: &mut CliTestContext) {
-        let name = ctx.wrap("test");
-        let out = cmd()
-            .args([
-                "-o",
-                "json",
-                "dataset",
-                "create",
-                "--region",
-                &ctx.region,
-                &name,
-            ])
-            .output()
-            .unwrap();
-        assert!(
-            out.status.success(),
-            "{}",
-            String::from_utf8_lossy(&out.stderr)
-        );
-        let result: CreateDatasetResult = out.json().unwrap();
-        assert_eq!(result.dataset.name, name);
-    }
-
-    #[test_context(CliTestContext)]
-    #[tokio::test]
-    async fn create_with_description(ctx: &mut CliTestContext) {
-        let name = ctx.wrap("test");
-        let out = cmd()
-            .args([
-                "-o",
-                "json",
-                "dataset",
-                "create",
-                "--region",
-                &ctx.region,
-                "--description",
-                "my dataset",
-                &name,
-            ])
-            .output()
-            .unwrap();
-        assert!(
-            out.status.success(),
-            "{}",
-            String::from_utf8_lossy(&out.stderr)
-        );
-        let result: CreateDatasetResult = out.json().unwrap();
-        assert_eq!(result.dataset.name, name);
-        assert_eq!(result.dataset.description.as_deref(), Some("my dataset"));
-    }
-
-    #[test_context(CliTestContext)]
-    #[tokio::test]
-    async fn get(ctx: &mut CliTestContext) {
-        let name = ctx.wrap("test");
-        cmd()
-            .args(["dataset", "create", "--region", &ctx.region, &name])
-            .output()
-            .unwrap();
-
-        let out = cmd()
-            .args(["-o", "json", "dataset", "get", &name])
-            .output()
-            .unwrap();
-        assert!(
-            out.status.success(),
-            "{}",
-            String::from_utf8_lossy(&out.stderr)
-        );
-        let result: GetDatasetResult = out.json().unwrap();
-        assert_eq!(result.dataset.name, name);
-    }
-
-    #[test_context(CliTestContext)]
-    #[tokio::test]
-    async fn update(ctx: &mut CliTestContext) {
-        let name = ctx.wrap("test");
-        cmd()
-            .args(["dataset", "create", "--region", &ctx.region, &name])
-            .output()
-            .unwrap();
-
-        let out = cmd()
-            .args([
-                "-o",
-                "json",
-                "dataset",
-                "update",
-                &name,
-                "--description",
-                "Hello world",
-            ])
-            .output()
-            .unwrap();
-        assert!(
-            out.status.success(),
-            "{}",
-            String::from_utf8_lossy(&out.stderr)
-        );
-        let result: UpdateDatasetResult = out.json().unwrap();
-        assert_eq!(result.dataset.name, name);
-        assert_eq!(result.dataset.description.as_deref(), Some("Hello world"));
-    }
-
-    #[test_context(CliTestContext)]
-    #[tokio::test]
-    async fn update_without_fields(ctx: &mut CliTestContext) {
-        let name = ctx.wrap("test");
-        cmd()
-            .args(["dataset", "create", "--region", &ctx.region, &name])
-            .output()
-            .unwrap();
-
-        let out = cmd().args(["dataset", "update", &name]).output().unwrap();
-        assert!(!out.status.success());
-        assert!(
-            String::from_utf8_lossy(&out.stderr).contains("at least one field must be specified")
-        );
-    }
-
-    #[test_context(CliTestContext)]
-    #[tokio::test]
-    async fn delete(ctx: &mut CliTestContext) {
-        let name = ctx.wrap("test");
-        cmd()
-            .args(["dataset", "create", "--region", &ctx.region, &name])
-            .output()
-            .unwrap();
-
-        let out = cmd()
-            .args(["-o", "json", "dataset", "delete", &name, "-y"])
-            .output()
-            .unwrap();
-        assert!(
-            out.status.success(),
-            "{}",
-            String::from_utf8_lossy(&out.stderr)
-        );
-        let result: DeleteDatasetResult = out.json().unwrap();
-        assert!(result.deleted);
-    }
-
-    #[test_context(CliTestContext)]
-    #[tokio::test]
-    async fn delete_aborted(ctx: &mut CliTestContext) {
-        let name = ctx.wrap("test");
-        cmd()
-            .args(["dataset", "create", "--region", &ctx.region, &name])
-            .output()
-            .unwrap();
-
-        let out = cmd()
-            .args(["-o", "json", "dataset", "delete", &name])
-            .write_stdin("wrong-name\n")
-            .output()
-            .unwrap();
-        assert!(out.status.success());
-        let result: DeleteDatasetResult = out.json().unwrap();
-        assert!(!result.deleted);
-    }
 }
