+++ topk-cli/src/lib.rs
@@ -1,4 +1,3 @@
-pub mod client;
 pub mod commands;
 pub mod config;
 pub mod dataset_region_cache;
